@@ -19,6 +19,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.samples.petclinic.api.dto.OwnerDetails;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
+import org.springframework.web.reactive.function.client.WebClient;
+import reactor.core.publisher.Mono;
 
 /**
  * @author Maciej Szarlinski
@@ -27,9 +29,12 @@ import org.springframework.web.client.RestTemplate;
 @RequiredArgsConstructor
 public class CustomersServiceClient {
 
-    private final RestTemplate loadBalancedRestTemplate;
+    private final WebClient.Builder webClientBuilder;
 
-    public OwnerDetails getOwner(final int ownerId) {
-        return loadBalancedRestTemplate.getForObject("http://customers-service/owners/{ownerId}", OwnerDetails.class, ownerId);
+    public Mono<OwnerDetails> getOwner(final int ownerId) {
+        return webClientBuilder.build().get()
+            .uri("http://customers-service/owners/{ownerId}", ownerId)
+            .retrieve()
+            .bodyToMono(OwnerDetails.class);
     }
 }
