@@ -15,22 +15,15 @@
  */
 package org.springframework.samples.petclinic.api.application;
 
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-
-import com.netflix.hystrix.contrib.javanica.annotation.HystrixCommand;
 import lombok.RequiredArgsConstructor;
-import org.springframework.samples.petclinic.api.dto.VisitDetails;
 import org.springframework.samples.petclinic.api.dto.Visits;
 import org.springframework.stereotype.Component;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.UriComponentsBuilder;
+import org.springframework.web.reactive.function.client.WebClient;
+import reactor.core.publisher.Mono;
 
+import java.util.List;
 
-import static java.util.stream.Collectors.groupingBy;
 import static java.util.stream.Collectors.joining;
-import static org.springframework.web.util.UriComponentsBuilder.fromHttpUrl;
 
 /**
  * @author Maciej Szarlinski
@@ -39,24 +32,24 @@ import static org.springframework.web.util.UriComponentsBuilder.fromHttpUrl;
 @RequiredArgsConstructor
 public class VisitsServiceClient {
 
-    private final RestTemplate loadBalancedRestTemplate;
+    // Could be changed for testing purpose
+    private String hostname = "http://visits-service/";
 
-    @HystrixCommand(fallbackMethod = "emptyVisitsForPets")
-    public Map<Integer, List<VisitDetails>> getVisitsForPets(final List<Integer> petIds) {
-        UriComponentsBuilder builder = fromHttpUrl("http://visits-service/pets/visits")
-            .queryParam("petId", joinIds(petIds));
+    private final WebClient.Builder webClientBuilder;
 
-        return loadBalancedRestTemplate.getForObject(builder.toUriString(), Visits.class)
-            .getItems()
-            .stream()
-            .collect(groupingBy(VisitDetails::getPetId));
+    public Mono<Visits> getVisitsForPets(final List<Integer> petIds) {
+        return webClientBuilder.build()
+            .get()
+            .uri(hostname + "pets/visits?petId={petId}", joinIds(petIds))
+            .retrieve()
+            .bodyToMono(Visits.class);
     }
 
     private String joinIds(List<Integer> petIds) {
         return petIds.stream().map(Object::toString).collect(joining(","));
     }
 
-    private Map<Integer, List<VisitDetails>> emptyVisitsForPets(List<Integer> petIds) {
-        return Collections.emptyMap();
+    void setHostname(String hostname) {
+        this.hostname = hostname;
     }
 }
