@@ -15,6 +15,7 @@
  */
 package org.springframework.samples.petclinic.visits.model;
 
+import java.util.Collection;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -33,5 +34,5 @@ public interface VisitRepository extends JpaRepository<Visit, Integer> {
 
     List<Visit> findByPetId(int petId);
 
-    List<Visit> findByPetIdIn(Iterable<Integer> petIds);
+    List<Visit> findByPetIdIn(Collection<Integer> petIds);
 }
